+++ RLBotPack/Necto/Nexto_Toxic/bot.cfg
@@ -1,12 +1,12 @@
 [Locations]
 # Path to loadout config. Can use relative path from here.
-looks_config = ./appearance.cfg
+looks_config = ../Nexto/appearance.cfg
 
 # Path to python file. Can use relative path from here.
 python_file = ./bot.py
-requirements_file = ./requirements.txt
+requirements_file = ../Nexto/requirements.txt
 
-logo_file = ./nexto_logo.png
+logo_file = ../Nexto/nexto_logo.png
 
 # Name of the bot in-game
 name = Nexto (Toxic!)
@@ -17,13 +17,13 @@ maximum_tick_rate_preference = 120
 [Details]
 # These values are optional but useful metadata for helper programs
 # Name of the bot's creator/developer
-developer = Rolv, Soren, and several contributors
+developer = The Collective Hatred of the Rocket League Community
 
 # Short description of the bot
-description = Nexto is version 2 of Necto, the official RLGym community bot, trained using PPO with workers run by people all around the world.
+description = Nexto, but angry
 
 # Fun fact about the bot
-fun_fact = Nexto uses an attention mechanism, commonly used for text understanding, to support any number of players
+fun_fact = No Fun Allowed
 
 # Link to github repository
 github = https://github.com/Rolv-Arild/Necto
